+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Clean command.
+- A simple example has been added.
+
 ## 1.0.0
 
 - Initial version.